Moved state managers in operators' modules
